restricting access to workout pages
to logged-in users

+++ frontend/src/components/WorkoutDetails.js
@@ -2,14 +2,24 @@ import React from 'react';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 import EditWorkout from '../components/EditWorkout'
 import { useWorkoutsContext } from '../hooks/useWorkoutContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 export default function WorkoutDetails(props){
     const [showEditForm, setShowEditForm] = React.useState(false);
+    const [error, setError] = React.useState(null);
     const {dispatch} = useWorkoutsContext();
-   console.log(`details ${props.id} rendered`)
+    const { user } = useAuthContext();
+    console.log(`details ${props.id} rendered`)
     const handleClick = ()=>{
+         if(!user){
+          setError('You must be logged in to do that')
+          return
+          }
         fetch('/api/workouts/' + props.id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         }).then(response => response.json())
         .then(json => dispatch({type: 'DELETE_ONE', payload: json }))
         .catch(err => console.log(err))
@@ -27,6 +37,7 @@ export default function WorkoutDetails(props){
             <p className='date'>{formatDistanceToNow(new Date(props.createdAt), { addSuffix: true})}</p>
             <span className='material-symbols-outlined' onClick={handleClick}>delete</span>
             <span className='material-symbols-outlined edit' onClick={showEdit}>edit</span>
+            {error && <div className="error">{error}</div>} 
         </div>
         {showEditForm && <EditWorkout
                         key={props.id +'edit'}
